Add files via upload
This project is only based of Linear Regression. Here, I've use a data set of a car company and took Mileage as the dependent and the rest as the independent variables to create a model. I've also tried to see if using dummy data can increase the accuracy and yup! it can. All the processes are documented in fine manner. I'm quite happy with it!